refactored removal of try-catch to improve performance
If string-to-ID conversion fails, now returning null, to indicate that
the request was bad, since id string was bad.  Note the Service layer
implementation now depends on the conversion to return null instead of
throwing Exception; we now don't _assume_ an exception could be thrown.
E.g. no exception catching is needed if the ID is of type String.